\name{Chi}
\docType{data}
\alias{Chi}
\alias{test2}
\alias{test3}
\alias{test4}

\title{ The demo dataset for \code{MiClip} package
}
\description{This demo dataset is a small portion of the single-end HITS-CLIP data provided in the Chi study.
}
\usage{data(Chi)}
\format{3 lists that are generated by different \code{MiClip} functions
}

\references{
  Chi, S.W., Zang, J.B., Mele, A. and Darnell, R.B. (2009) Argonaute HITS-CLIP decodes microRNA-mRNA interaction maps. Nature, 460, 479-486.
}

\keyword{datasets}
